requirements.txt
python-telegram-bot
streamlit
openai
pandas
python-dotenv
fastapi
pydantic
uvicorn[standard]
PyPDF2
